requirements.txt
aiohttp
matplotlib
loguru
pandas
requests
streamlit
aiogram
python-dotenv
